Add common unit test utils as a separate project
Extract utils common to all unit tests as a separate project so it can
be references by any unit test project requiring it.
